backend/requirements.txt
# FastAPI Backend Dependencies

# Core Framework
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Email
fastapi-mail==1.4.1

# File Storage
boto3==1.34.19  # For AWS S3
minio==7.2.3    # For local MinIO

# Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
ruff==0.1.14

# Utilities
python-slugify==8.0.1
pillow==10.2.0
